upd: remove CblockFileInfo from main
