+++ divi/src/main.cpp
@@ -38,6 +38,7 @@
 #include <BlockUndo.h>
 #include <ValidationState.h>
 #include <scriptCheck.h>
+#include <blockFileInfo.h>
 
 using namespace boost;
 using namespace std;
@@ -5621,13 +5622,6 @@ bool SendMessages(CNode* pto, bool fSendTrickle)
     }
     return true;
 }
-
-std::string CBlockFileInfo::ToString() const
-{
-    return strprintf("CBlockFileInfo(blocks=%u, size=%u, heights=%u...%u, time=%s...%s)", nBlocks, nSize, nHeightFirst, nHeightLast, DateTimeStrFormat("%Y-%m-%d", nTimeFirst), DateTimeStrFormat("%Y-%m-%d", nTimeLast));
-}
-
-
 class CMainCleanup
 {
 public:
